refactoring Menu
