+++ src/main/java/br/upe/controller/ui/HelloController.java
@@ -1,8 +1,5 @@
 package br.upe.controller.ui;
 
-import javafx.fxml.FXML;
-import javafx.scene.control.Label;
-
 public class HelloController {
         @FXML
         private Label welcomeText;
