+++ src/main/java/br/upe/ui/MenuExercicios.java
@@ -1,8 +1,8 @@
 package br.upe.ui;
 
-import br.upe.business.ExercicioService;
+import br.upe.controller.business.ExercicioService;
 import br.upe.data.beans.Exercicio;
-import br.upe.business.IExercicioService;
+import br.upe.controller.business.IExercicioService;
 
 import java.util.List;
 import java.util.Optional;
